# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "AI Mafia Backend",
        "description": "Backend для игры в мафию с AI игроками",
        "version": "1.0.0"
    },
    "paths": {
        "/api/rooms/": {
            "get": {
                "tags": [
                    "rooms"
                ],
                "summary": "List Rooms",
                "description": "Получить список активных лобби для браузера комнат.\nВозвращает только комнаты со статусом 'lobby' или 'starting'.",
                "operationId": "list_rooms_api_rooms__get",
                "parameters": [
                    {
                        "name": "skip",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "title": "Skip"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20,
                            "title": "Limit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Room"
                                    },
                                    "title": "Response List Rooms Api Rooms  Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "rooms"
                ],
                "summary": "Create Room",
                "description": "Создать новую комнату.",
                "operationId": "create_room_api_rooms__post",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RoomCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Room"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/rooms/{room_id}": {
            "get": {
                "tags": [
                    "rooms"
                ],
                "summary": "Get Room",
                "description": "Получить комнату по её публичному room_id (UUID).",
                "operationId": "get_room_api_rooms__room_id__get",
                "parameters": [
                    {
                        "name": "room_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Room Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Room"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "rooms"
                ],
                "summary": "Update Room",
                "description": "Обновить настройки комнаты.\nТолько хост может менять настройки (проверка host_token через заголовок при необходимости).",
                "operationId": "update_room_api_rooms__room_id__patch",
                "parameters": [
                    {
                        "name": "room_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Room Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RoomUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Room"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "rooms"
                ],
                "summary": "Delete Room",
                "description": "Удалить комнату и уведомить подключённых игроков.",
                "operationId": "delete_room_api_rooms__room_id__delete",
                "parameters": [
                    {
                        "name": "room_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Room Id"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/rooms/{room_id}/join": {
            "post": {
                "tags": [
                    "rooms"
                ],
                "summary": "Join Room",
                "description": "Присоединиться к комнате как игрок.",
                "operationId": "join_room_api_rooms__room_id__join_post",
                "parameters": [
                    {
                        "name": "room_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Room Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PlayerCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Player"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/rooms/{room_id}/players": {
            "get": {
                "tags": [
                    "rooms"
                ],
                "summary": "Get Room Players",
                "description": "Получить всех игроков в комнате.",
                "operationId": "get_room_players_api_rooms__room_id__players_get",
                "parameters": [
                    {
                        "name": "room_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Room Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Player"
                                    },
                                    "title": "Response Get Room Players Api Rooms  Room Id  Players Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/rooms/{room_id}/game": {
            "get": {
                "tags": [
                    "rooms"
                ],
                "summary": "Get Room Game",
                "description": "Получить текущую игровую сессию для комнаты.",
                "operationId": "get_room_game_api_rooms__room_id__game_get",
                "parameters": [
                    {
                        "name": "room_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Room Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Game"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/rooms/{room_id}/game/state": {
            "get": {
                "tags": [
                    "rooms"
                ],
                "summary": "Get Game State",
                "description": "Получить текущую фазу игры, номер дня, список живых игроков.",
                "operationId": "get_game_state_api_rooms__room_id__game_state_get",
                "parameters": [
                    {
                        "name": "room_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Room Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": true,
                                    "title": "Response Get Game State Api Rooms  Room Id  Game State Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/rooms/{room_id}/game/start": {
            "post": {
                "tags": [
                    "rooms"
                ],
                "summary": "Start Game",
                "description": "HTTP-альтернатива WebSocket start_game событию.\nЗапускает игру в комнате.",
                "operationId": "start_game_api_rooms__room_id__game_start_post",
                "parameters": [
                    {
                        "name": "room_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Room Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "additionalProperties": true,
                                    "title": "Response Start Game Api Rooms  Room Id  Game Start Post"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/rooms/{room_id}/game/events": {
            "get": {
                "tags": [
                    "rooms"
                ],
                "summary": "Get Game Events",
                "description": "Получить историю событий текущей игры в комнате.",
                "operationId": "get_game_events_api_rooms__room_id__game_events_get",
                "parameters": [
                    {
                        "name": "room_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Room Id"
                        }
                    },
                    {
                        "name": "skip",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "title": "Skip"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 100,
                            "title": "Limit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/GameEvent"
                                    },
                                    "title": "Response Get Game Events Api Rooms  Room Id  Game Events Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/rooms/{room_id}/players/{player_id}": {
            "delete": {
                "tags": [
                    "rooms"
                ],
                "summary": "Kick Player",
                "description": "Удалить игрока из комнаты (кик через HTTP).",
                "operationId": "kick_player_api_rooms__room_id__players__player_id__delete",
                "parameters": [
                    {
                        "name": "room_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Room Id"
                        }
                    },
                    {
                        "name": "player_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Player Id"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful Response"
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "PlayerCreate": {
                "properties": {
                    "player_id": {
                        "type": "string",
                        "title": "Player Id"
                    },
                    "room_id": {
                        "type": "integer",
                        "title": "Room Id"
                    },
                    "nickname": {
                        "type": "string",
                        "title": "Nickname"
                    },
                    "is_ai": {
                        "type": "boolean",
                        "title": "Is Ai",
                        "default": false
                    },
                    "role": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Role"
                    },
                    "is_alive": {
                        "type": "boolean",
                        "title": "Is Alive",
                        "default": true
                    },
                    "is_connected": {
                        "type": "boolean",
                        "title": "Is Connected",
                        "default": true
                    },
                    "session_token": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Session Token"
                    }
                },
                "type": "object",
                "required": [
                    "player_id",
                    "room_id",
                    "nickname"
                ],
                "title": "PlayerCreate"
            },
...
        }
    }
}